+++ src/commands/clear_command.py
@@ -2,6 +2,9 @@ import os
 from src.commands.base_command import BaseCommand
 
 class ClearCommand(BaseCommand):
+    def __init__(self, api_client, console, cache):
+        super().__init__(api_client, console, cache)
+        
     def get_help_message(self):
         return {
             "description": "Nettoie l'écran du terminal.",